Transfer WithAssumptions from core.py to assumptions.py
